test: Separating pattern analysis tests by source graph
